powerjoin: Extensions of 'dplyr' and 'fuzzyjoin' Join Functions

We extend 'dplyr' and 'fuzzyjoin' join functions with features to preprocess the data, apply various data checks, and deal with conflicting columns.

See also

Author

Maintainer: Antoine Fabri antoine.fabri@gmail.com

Other contributors:

Hadley Wickham hadley@rstudio.com (ORCID) (aut/cre of dplyr) [contributor]

Romain François (ORCID) (aut of dplyr) [contributor]

David Robinson admiral.david@gmail.com (aut of fuzzyjoin) [contributor]

RStudio (cph/fnd dplyr) [copyright holder, funder]